# Migrate to stackdriver trace API v2.