--- a/sender-stackdriver/src/main/java/zipkin2/reporter/stackdriver/StackdriverSender.java
+++ b/sender-stackdriver/src/main/java/zipkin2/reporter/stackdriver/StackdriverSender.java
@@ -201,7 +201,7 @@ public final class StackdriverSender extends Sender {
 
     @Override
     public String toString() {
-      return "PatchTracesCall{" + request() + "}";
+      return "BatchWriteSpansCall{" + request() + "}";
     }
 
     @Override