--- a/translation-stackdriver/src/main/java/zipkin2/translation/stackdriver/SpanTranslator.java
+++ b/translation-stackdriver/src/main/java/zipkin2/translation/stackdriver/SpanTranslator.java
@@ -100,7 +100,7 @@ public final class SpanTranslator {
     return spanBuilder;
   }
 
-  private static Timestamp createTimestamp(long microseconds) {
+  static Timestamp createTimestamp(long microseconds) {
     long seconds = (microseconds / 1000000);
     int remainderMicros = (int) (microseconds % 1000000);
     int remainderNanos = remainderMicros * 1000;