Took a stab at the migration. Not having to group spans into traces means things become much much simpler. Allocation-wise, since we have to combine trace ID and span ID into a single field for every span now, I think this is the best we can do.

Fixes #111 

Benchmarks. I used a modified benchmark of the parser that serializes the proto since some of the optimizations take advantage of the expectation that the proto will be serialized at some point (using `ByteString` as much as possible, `getSpanIdBytes()`, `UnsafeByteOperations.wrap`). It's still not quite as fast as before, but don't think it would be able to match it considering now a new array must be allocated for every span which concatenates trace ID and span ID and the truncatable strings / attribute values probably add some overhead. As sort of expected, I do notice that the serializing version of the benchmark is ~2x slower on the old code while only ~60% slower with the new one.

New

```
Benchmark                                                                   Mode  Cnt  Score   Error  Units
StackdriverEncoderBenchmarks.encodeClientSpan_json_stackdriver_proto3       avgt   15  1.739 ± 0.037  us/op
StackdriverEncoderBenchmarks.encodeClientSpan_json_zipkin_json              avgt   15  0.755 ± 0.007  us/op
StackdriverEncoderBenchmarks.sizeInBytesClientSpan_json_stackdriver_proto3  avgt   15  0.976 ± 0.027  us/op
StackdriverEncoderBenchmarks.sizeInBytesClientSpan_json_zipkin_json         avgt   15  0.118 ± 0.004  us/op

Benchmark                                   Mode  Cnt    Score   Error  Units
TracesParserBenchmarks.parse100ClientSpans  avgt   15  258.795 ± 3.177  us/op
TracesParserBenchmarks.parseClientSpan      avgt   15    2.579 ± 0.024  us/op
```

Old

```
Benchmark                                                                   Mode  Cnt  Score   Error  Units
StackdriverEncoderBenchmarks.encodeClientSpan_json_stackdriver_proto3       avgt   15  1.796 ± 0.049  us/op
StackdriverEncoderBenchmarks.encodeClientSpan_json_zipkin_json              avgt   15  0.797 ± 0.035  us/op
StackdriverEncoderBenchmarks.sizeInBytesClientSpan_json_stackdriver_proto3  avgt   15  1.162 ± 0.024  us/op
StackdriverEncoderBenchmarks.sizeInBytesClientSpan_json_zipkin_json         avgt   15  0.116 ± 0.001  us/op

Benchmark                                   Mode  Cnt    Score   Error  Units
TracesParserBenchmarks.parse100ClientSpans  avgt   15  157.596 ± 2.184  us/op
TracesParserBenchmarks.parseClientSpan      avgt   15    1.770 ± 0.044  us/op
```

/cc @ewhauser