@ewhauser 2018-12-31
This is good to be merged from my perspective.

I tracked down the issue with the tags. It's a bug in the Zipkin JS library which is sending tags as binaryAnnotations to the v2 endpoint. I'll open an issue over there.